critical battery warning in offline Mode
Batterywarning in Offline mode

batterywarning is triggered in offline mode when the batttery is below the batt_min value retrieved from the eeprom during startup. The device gives the user a chance to set correct parameters in the options menu, in case they are faulty. If no changes are made the drevice shuts down